Moves functional test from integration test suite to unit tests
