+++ services/s3/src/test/java/software/amazon/awssdk/services/s3/functionaltests/AsyncResponseTransformerTest.java
@@ -13,13 +13,14 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.services.s3;
+package software.amazon.awssdk.services.s3.functionaltests;
 
 import static org.assertj.core.api.Assertions.fail;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
+import java.util.concurrent.CompletionException;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
@@ -27,15 +28,16 @@ import org.mockito.runners.MockitoJUnitRunner;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProviderChain;
 import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
 import software.amazon.awssdk.core.async.AsyncResponseTransformer;
+import software.amazon.awssdk.services.s3.S3AsyncClient;
 
 @RunWith(MockitoJUnitRunner.class)
-public class AsyncResponseTransformerIntegrationTest {
+public class AsyncResponseTransformerTest {
 
     @Mock
     private AsyncResponseTransformer asyncResponseTransformer;
 
     @Test
-    public void AsyncResponseTransformerPrepareCalled_BeforeCredentailsResolution() {
+    public void AsyncResponseTransformerPrepareCalled_BeforeCredentialsResolution() {
         S3AsyncClient client = S3AsyncClient.builder()
                                             .credentialsProvider(AwsCredentialsProviderChain.of(
                                                 ProfileCredentialsProvider.create("dummyprofile")))
@@ -44,8 +46,7 @@ public class AsyncResponseTransformerIntegrationTest {
         try {
             client.getObject(b -> b.bucket("dummy").key("key"), asyncResponseTransformer).join();
             fail("Expected an exception during credential resolution");
-        } catch (Throwable t) {
-
+        } catch (CompletionException e) {
         }
 
         verify(asyncResponseTransformer, times(1)).prepare();
